requirements.txt
numpy==1.23.2
pandas==1.5.3
python-dotenv==1.0.1
datetime==5.4
pytz==2024.1
unidecode==1.3.8
requests==2.31.0
